+++ smart-system-core/src/main/java/org/jeecg/modules/SmartInnerPartyTalk/entity/SmartInnerPartyPacpa.java
@@ -1,0 +1,58 @@
+package org.jeecg.modules.SmartInnerPartyTalk.entity;
+
+import java.io.Serializable;
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
+import lombok.Data;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import org.springframework.format.annotation.DateTimeFormat;
+import org.jeecgframework.poi.excel.annotation.Excel;
+import java.util.Date;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+import java.io.UnsupportedEncodingException;
+
+/**
+ * @Description: 党内谈话参与人表
+ * @Author: jeecg-boot
+ * @Date:   2021-11-05
+ * @Version: V1.0
+ */
+@ApiModel(value="smart_inner_party_pacpa对象", description="党内谈话参与人表")
+@Data
+@TableName("smart_inner_party_pacpa")
+public class SmartInnerPartyPacpa implements Serializable {
+    private static final long serialVersionUID = 1L;
+
+	/**主键*/
+	@TableId(type = IdType.ASSIGN_ID)
+    @ApiModelProperty(value = "主键")
+    private java.lang.String id;
+	/**创建人*/
+    @ApiModelProperty(value = "创建人")
+    private java.lang.String createBy;
+	/**创建日期*/
+	@JsonFormat(timezone = "GMT+8",pattern = "yyyy-MM-dd HH:mm:ss")
+    @DateTimeFormat(pattern="yyyy-MM-dd HH:mm:ss")
+    @ApiModelProperty(value = "创建日期")
+    private java.util.Date createTime;
+	/**更新人*/
+    @ApiModelProperty(value = "更新人")
+    private java.lang.String updateBy;
+	/**更新日期*/
+	@JsonFormat(timezone = "GMT+8",pattern = "yyyy-MM-dd HH:mm:ss")
+    @DateTimeFormat(pattern="yyyy-MM-dd HH:mm:ss")
+    @ApiModelProperty(value = "更新日期")
+    private java.util.Date updateTime;
+	/**所属部门*/
+    @ApiModelProperty(value = "所属部门")
+    private java.lang.String sysOrgCode;
+	/**主表ID*/
+    @ApiModelProperty(value = "主表ID")
+    private java.lang.String parentId;
+	/**参会人员工号*/
+	@Excel(name = "参会人员工号", width = 15)
+    @ApiModelProperty(value = "参会人员工号")
+    private java.lang.String papcNo;
+}
