+++ smart-system-core/src/main/java/org/jeecg/modules/SmartInnerPartyTalk/mapper/SmartInnerPartyTalkMapper.java
@@ -1,0 +1,17 @@
+package org.jeecg.modules.SmartInnerPartyTalk.mapper;
+
+import java.util.List;
+
+import org.apache.ibatis.annotations.Param;
+import org.jeecg.modules.SmartInnerPartyTalk.entity.SmartInnerPartyTalk;
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+
+/**
+ * @Description: 党内谈话表
+ * @Author: jeecg-boot
+ * @Date:   2021-11-05
+ * @Version: V1.0
+ */
+public interface SmartInnerPartyTalkMapper extends BaseMapper<SmartInnerPartyTalk> {
+
+}
