+++ smart-system-core/src/main/java/org/jeecg/modules/SmartInnerPartyTalk/mapper/xml/SmartInnerPartyAnnexMapper.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="org.jeecg.modules.SmartInnerPartyTalk.mapper.SmartInnerPartyAnnexMapper">
+
+	<delete id="deleteByMainId" parameterType="java.lang.String">
+		DELETE 
+		FROM  smart_inner_party_annex 
+		WHERE
+			 parent_id = #{mainId} 	</delete>
+	
+	<select id="selectByMainId" parameterType="java.lang.String" resultType="org.jeecg.modules.SmartInnerPartyTalk.entity.SmartInnerPartyAnnex">
+		SELECT * 
+		FROM  smart_inner_party_annex
+		WHERE
+			 parent_id = #{mainId} 	</select>
+</mapper>
