--- smart-system-core/src/main/java/org/jeecg/modules/smartInnerParty/mapper/SmartInnerPartyTalkMapper.java
@@ -1,17 +1,0 @@
-package org.jeecg.modules.smartInnerParty.mapper;
-
-import java.util.List;
-
-import org.apache.ibatis.annotations.Param;
-import org.jeecg.modules.smartInnerParty.entity.SmartInnerPartyTalk;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-
-/**
- * @Description: 党内谈话表
- * @Author: jeecg-boot
- * @Date:   2021-11-02
- * @Version: V1.0
- */
-public interface SmartInnerPartyTalkMapper extends BaseMapper<SmartInnerPartyTalk> {
-
-}
